refactor(timeline): move the crypto tasks spawning to `decryption_retry_task` too
This will make it easier to move all the code around, when needs be.
