+++ crates/matrix-sdk-ui/src/timeline/controller/decryption_retry_task.rs
@@ -12,37 +12,48 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{collections::BTreeSet, sync::Arc};
+use std::{
+    collections::{BTreeMap, BTreeSet},
+    sync::Arc,
+};
 
+use futures_core::Stream;
+use futures_util::pin_mut;
 use imbl::Vector;
 use itertools::{Either, Itertools as _};
 use matrix_sdk::{
+    crypto::store::types::RoomKeyInfo,
     deserialized_responses::TimelineEventKind as SdkTimelineEventKind,
-    event_handler::EventHandlerHandle, executor::JoinHandle, Client,
+    encryption::backups::BackupState,
+    event_handler::EventHandlerHandle,
+    executor::{spawn, JoinHandle},
+    Client, Room,
 };
 use tokio::sync::{
     mpsc::{self, Receiver, Sender},
     RwLock,
 };
-use tracing::{debug, error, field, info, info_span, Instrument as _};
+use tokio_stream::{wrappers::errors::BroadcastStreamRecvError, StreamExt as _};
+use tracing::{debug, error, field, info, info_span, warn, Instrument as _};
 
 use crate::timeline::{
     controller::{TimelineSettings, TimelineState},
     event_item::EventTimelineItemKind,
+    to_device::{handle_forwarded_room_key_event, handle_room_key_event},
     traits::{Decryptor, RoomDataProvider},
-    EncryptedMessage, EventTimelineItem, TimelineItem, TimelineItemKind,
+    EncryptedMessage, EventTimelineItem, TimelineController, TimelineItem, TimelineItemKind,
 };
 
 /// All the drop handles for the tasks used for crypto, namely message
 /// re-decryption, in the timeline.
 #[derive(Debug)]
 pub(in crate::timeline) struct CryptoDropHandles {
-    pub client: Client,
-    pub event_handler_handles: Vec<EventHandlerHandle>,
-    pub room_key_from_backups_join_handle: JoinHandle<()>,
-    pub room_keys_received_join_handle: JoinHandle<()>,
-    pub room_key_backup_enabled_join_handle: JoinHandle<()>,
-    pub encryption_changes_handle: JoinHandle<()>,
+    client: Client,
+    event_handler_handles: Vec<EventHandlerHandle>,
+    room_key_from_backups_join_handle: JoinHandle<()>,
+    room_keys_received_join_handle: JoinHandle<()>,
+    room_key_backup_enabled_join_handle: JoinHandle<()>,
+    encryption_changes_handle: JoinHandle<()>,
 }
 
 impl Drop for CryptoDropHandles {
@@ -55,6 +66,158 @@ impl Drop for CryptoDropHandles {
         self.room_keys_received_join_handle.abort();
         self.room_key_backup_enabled_join_handle.abort();
         self.encryption_changes_handle.abort();
+    }
+}
+
+/// The task that handles the room keys from backups.
+async fn room_keys_from_backups_task<S>(stream: S, timeline_controller: TimelineController)
+where
+    S: Stream<Item = Result<BTreeMap<String, BTreeSet<String>>, BroadcastStreamRecvError>>,
+{
+    pin_mut!(stream);
+
+    while let Some(update) = stream.next().await {
+        match update {
+            Ok(info) => {
+                let mut session_ids = BTreeSet::new();
+
+                for set in info.into_values() {
+                    session_ids.extend(set);
+                }
+
+                timeline_controller.retry_event_decryption(Some(session_ids)).await;
+            }
+            // We lagged, so retry every event.
+            Err(_) => timeline_controller.retry_event_decryption(None).await,
+        }
+    }
+}
+
+/// The task that handles the [`BackupState`] updates.
+async fn backup_states_task<S>(backup_states_stream: S, timeline_controller: TimelineController)
+where
+    S: Stream<Item = Result<BackupState, BroadcastStreamRecvError>>,
+{
+    pin_mut!(backup_states_stream);
+
+    while let Some(update) = backup_states_stream.next().await {
+        match update {
+            // If the backup got enabled, or we lagged and thus missed that the backup
+            // might be enabled, retry to decrypt all the events. Please note, depending
+            // on the backup download strategy, this might do two things under the
+            // assumption that the backup contains the relevant room keys:
+            //
+            // 1. It will decrypt the events, if `BackupDownloadStrategy` has been set to `OneShot`.
+            // 2. It will fail to decrypt the event, but try to download the room key to decrypt it
+            //    if the `BackupDownloadStrategy` has been set to `AfterDecryptionFailure`.
+            Ok(BackupState::Enabled) | Err(_) => {
+                timeline_controller.retry_event_decryption(None).await;
+            }
+            // The other states aren't interesting since they are either still enabling
+            // the backup or have the backup in the disabled state.
+            Ok(
+                BackupState::Unknown
+                | BackupState::Creating
+                | BackupState::Resuming
+                | BackupState::Disabling
+                | BackupState::Downloading
+                | BackupState::Enabling,
+            ) => (),
+        }
+    }
+}
+
+/// The task that handles the [`RoomKeyInfo`] updates.
+async fn room_key_received_task<S>(
+    room_keys_received_stream: S,
+    timeline_controller: TimelineController,
+) where
+    S: Stream<Item = Result<Vec<RoomKeyInfo>, BroadcastStreamRecvError>>,
+{
+    pin_mut!(room_keys_received_stream);
+
+    let room_id = timeline_controller.room().room_id();
+
+    while let Some(room_keys) = room_keys_received_stream.next().await {
+        let session_ids = match room_keys {
+            Ok(room_keys) => {
+                let session_ids: BTreeSet<String> = room_keys
+                    .into_iter()
+                    .filter(|info| info.room_id == room_id)
+                    .map(|info| info.session_id)
+                    .collect();
+
+                Some(session_ids)
+            }
+            Err(BroadcastStreamRecvError::Lagged(missed_updates)) => {
+                // We lagged, let's retry to decrypt anything we have, maybe something
+                // was received.
+                warn!(
+                    missed_updates,
+                    "The room keys stream has lagged, retrying to decrypt the whole timeline"
+                );
+
+                None
+            }
+        };
+
+        timeline_controller.retry_event_decryption(session_ids).await;
+    }
+}
+
+/// Spawn all the crypto-related tasks that are used to handle re-decryption of
+/// messages.
+pub(in crate::timeline) async fn spawn_crypto_tasks(
+    room: Room,
+    controller: TimelineController,
+) -> CryptoDropHandles {
+    let room_key_handle = room
+        .client()
+        .add_event_handler(handle_room_key_event(controller.clone(), room.room_id().to_owned()));
+
+    let client = room.client();
+    let forwarded_room_key_handle = client.add_event_handler(handle_forwarded_room_key_event(
+        controller.clone(),
+        room.room_id().to_owned(),
+    ));
+
+    let event_handlers = vec![room_key_handle, forwarded_room_key_handle];
+
+    // Not using room.add_event_handler here because RoomKey events are
+    // to-device events that are not received in the context of a room.
+
+    let room_key_from_backups_join_handle = spawn(room_keys_from_backups_task(
+        client.encryption().backups().room_keys_for_room_stream(controller.room().room_id()),
+        controller.clone(),
+    ));
+
+    let room_key_backup_enabled_join_handle =
+        spawn(backup_states_task(client.encryption().backups().state_stream(), controller.clone()));
+
+    // TODO: Technically, this should be the only stream we need to listen to get
+    // notified when we should retry to decrypt an event. We sadly can't do that,
+    // since the cross-process support kills the `OlmMachine` which then in
+    // turn kills this stream. Once this is solved remove all the other ways we
+    // listen for room keys.
+    let room_keys_received_join_handle = {
+        spawn(room_key_received_task(
+            client.encryption().room_keys_received_stream().await.expect(
+                "We should be logged in by now, so we should have access to an `OlmMachine` \
+                     to be able to listen to this stream",
+            ),
+            controller.clone(),
+        ))
+    };
+
+    CryptoDropHandles {
+        client,
+        event_handler_handles: event_handlers,
+        room_key_from_backups_join_handle,
+        room_keys_received_join_handle,
+        room_key_backup_enabled_join_handle,
+        encryption_changes_handle: spawn(async move {
+            controller.handle_encryption_state_changes().await
+        }),
     }
 }
 
@@ -96,8 +259,7 @@ impl<P: RoomDataProvider, D: Decryptor> DecryptionRetryTask<P, D> {
 
         // Spawn the long-running task, providing the receiver so we can listen for
         // decryption requests
-        let handle =
-            matrix_sdk::executor::spawn(decryption_task(state, room_data_provider, receiver));
+        let handle = spawn(decryption_task(state, room_data_provider, receiver));
 
         // Keep hold of the sender so we can send off decryption requests to the task.
         Self { sender, _task_handle: Arc::new(handle), _phantom: Default::default() }
