+++ crates/matrix-sdk-ui/src/timeline/controller/mod.rs
@@ -90,7 +90,7 @@ mod state;
 mod state_transaction;
 
 pub(super) use aggregations::*;
-pub(super) use decryption_retry_task::CryptoDropHandles;
+pub(super) use decryption_retry_task::{spawn_crypto_tasks, CryptoDropHandles};
 
 /// Data associated to the current timeline focus.
 ///
